+++ Assets/Code/Gameplay/Combat/DamageHandlers/DamageHandlerBase.cs
