+++ Assets/Code/Gameplay/Enemies/Base/EnemyMovementBase.cs
@@ -1,7 +1,4 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
-using TMPro;
 using UnityEngine;
 using Zenject;
