+++ Assets/Code/Gameplay/Enemies/Base/StateMachine/EnemyGuardStateBase.cs
@@ -1,7 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 namespace Game.Room.Enemy
 {
     public abstract class EnemyGuardStateBase : EnemyStateBase
