+++ Assets/Code/Gameplay/Enemies/Base/StateMachine/EnemyStateBase.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using Zenject;
