--- Assets/Code/Gameplay/Enemies/DEPRECATED_EnemyMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40c13cfa519b27d4abcc2dfedffdc637
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
