--- Assets/Code/Gameplay/Enemies/DEPRECATED_EnemyState.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace Game.Room.Enemy
-{
-    public enum DEPRECATED_EnemyState
-    {
-        Normal = 0,
-        Combat = 1,
-        Defeat = 2,
-    }
-}
