--- Assets/Code/Gameplay/Enemies/DEPRECATED_EnemyState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 683fe811c94dbf74bbd084e8e59b8881
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
