--- Assets/Code/Gameplay/Enemies/Derived/DEPRECATED_TestEnemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a9c5731b60fa0af4cad4b9bf05371a86
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
