+++ Assets/Code/Gameplay/Enemies/Derived/TestEnemy/TestEnemy.cs
@@ -1,6 +1,4 @@
 using Game.Combat;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace Game.Room.Enemy
