+++ Assets/Code/Gameplay/Enemies/Derived/TestEnemy/TestEnemyCombatState.cs
@@ -1,7 +1,4 @@
 using Game.Management;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 using Zenject;
 
 namespace Game.Room.Enemy
