+++ Assets/Code/Gameplay/Enemies/Derived/TestEnemy/TestEnemyDefeatedState.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace Game.Room.Enemy
