+++ Assets/Code/Gameplay/Enemies/Management/EnemyInstaller.cs
@@ -1,6 +1,4 @@
 using Game.Combat;
-using Game.Player.Ship;
-using System.Collections.Generic;
 using UnityEngine;
 using Zenject;
