+++ Assets/Code/Gameplay/Enemies/Management/EnemyManager.cs
@@ -1,6 +1,5 @@
 using Game.Management;
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using Zenject;
