+++ Assets/Code/Gameplay/Player/Enviro/Secrets/A10GunToPickUp.cs
@@ -1,6 +1,4 @@
 using Game.Player.Ship;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
