+++ Assets/Code/Gameplay/Player/Ship/ShipParts/Gun/Base/GunBase.cs
@@ -1,9 +1,3 @@
-using Game.Player;
-using Game.Player.Ship;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 namespace Game.Player.Ship
 {
     public abstract class GunBase : ShipPart, IGun
