+++ Assets/Code/Gameplay/Player/Ship/ShipParts/Gun/Module/Derived/Bullet.cs
@@ -1,5 +1,5 @@
+using Game.Combat;
 using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace Game.Player.Ship
