+++ Assets/Code/Gameplay/Player/Ship/ShipParts/Gun/Module/Derived/LaserGun.cs
@@ -1,7 +1,4 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using Zenject;
 
 namespace Game.Player.Ship
 {
