+++ Assets/Code/Gameplay/Player/Ship/ShipParts/Gun/Module/Derived/MachineGun.cs
@@ -1,8 +1,4 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEditor.Rendering;
 using UnityEngine;
-using Zenject;
 
 namespace Game.Player.Ship
 {
