+++ Assets/Code/Gameplay/Player/Ship/ShipParts/Gun/Module/Derived/Rocket.cs
@@ -1,3 +1,4 @@
+using Game.Combat;
 using System.Collections;
 using UnityEngine;
