+++ Assets/Code/Gameplay/Player/Ship/ShipParts/Gun/Module/Derived/RocketLuncher.cs
@@ -1,6 +1,4 @@
 using UnityEngine;
-using Zenject;
-using Game.Input.System;
 
 namespace Game.Player.Ship
 {
