+++ Assets/Prefabs/Enemies/EnemiesVariants/TestEnemy.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 0afc8354f20aa634388a322a9525e79a
-PrefabImporter:
+guid: c5f0daf1b8bf19f4f8d1d452f93af1a7
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
